pyproject.toml
[tool.poetry]
name = "seeya-server"
version = "0.1.0"
description = ""
authors = ["최병욱 <esc5221@gmail.com>", "Minkyu Lee <whizkyu@snu.ac.kr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.2"
django-extensions = "^3.2.3"
django-ninja = "^1.1.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
boto3 = "^1.34.42"
requests = "^2.31.0"
pillow = "^10.2.0"
django-storages = "^1.14.2"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
